chore: reorganize docs
