+++ docs/snippets/composable/Config.vue
