+++ docs/snippets/composable/LabelsLocalization.vue
