+++ docs/snippets/vuetify/Config.vue
