+++ docs/snippets/vuetify/CountryDisplaySlot.vue
