+++ docs/snippets/vuetify/Slots.vue
