+++ tsconfig.json
@@ -1,7 +1,9 @@
 {
   "extends": "@vue/tsconfig/tsconfig.dom.json",
   "include": ["src/**/*", "src/**/*.vue", "docs/**/*", "docs/**/*.vue"],
-  "exclude": ["docs/guide/snippets/**/*"],
+  "exclude": [
+    "docs/snippets/**/*"
+  ],
   "compilerOptions": {
     "tsBuildInfoFile": "./node_modules/.tmp/tsconfig.app.tsbuildinfo",
     "types": [
